OpenVINO multi version support: C++ support
