+++ onnxruntime/core/providers/openvino/backends/basic_backend.cc
@@ -87,8 +87,13 @@ void BasicBackend::StartAsyncInference(Ort::CustomOpApi& ort,
     auto graph_input_buffer = graph_input_blob->buffer()
                                   .as<InferenceEngine::PrecisionTrait<InferenceEngine::Precision::FP32>::value_type*>();
     size_t input_data_size = graph_input_blob->byteSize();
-    
+
+    #if (defined OPENVINO_2020_2) || (defined OPENVINO_2020_3)
+    const OrtValue* tensor = ort.KernelContext_GetInput(context, subgraph_context_.input_indexes[i]);
+    #else
     const OrtValue* tensor = ort.KernelContext_GetInput(context, subgraph_context_.input_names.at(input_info_iter->first));
+    #endif
+
     auto tensor_shape = ort.GetTensorTypeAndShape(tensor);
     auto elem_type = ort.GetTensorElementType(tensor_shape);
