+++ onnxruntime/core/providers/openvino/contexts.h
@@ -22,7 +22,11 @@ struct SubGraphContext {
   bool enable_batching = false;
   bool set_vpu_config = false;
   std::string subgraph_name;
+  #if (defined OPENVINO_2020_2) || (defined OPENVINO_2020_3)
+  std::vector<int> input_indexes;
+  #else
   std::unordered_map<std::string, int> input_names;
+  #endif
   std::unordered_map<std::string, int> output_names;
   std::string device_id;
   InferenceEngine::Precision precision;
