+++ onnxruntime/core/providers/openvino/ov_versions/capabilities.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "core/framework/compute_capability.h"
+
+namespace onnxruntime {
+namespace openvino_ep {
+
+#if (defined OPENVINO_2020_2) || (defined OPENVINO_2020_3)
+std::vector<std::unique_ptr<ComputeCapability>>
+GetCapability_2020_2(const onnxruntime::GraphViewer& graph_viewer, const std::string device_id);
+
+#elif defined OPENVINO_2020_4
+std::vector<std::unique_ptr<ComputeCapability>>
+GetCapability_2020_4(const onnxruntime::GraphViewer& graph_viewer, const std::string device_id);
+
+#endif
+
+} //namespace openvino_ep
+} //namespace onnxruntime
